requirements.txt
python-telegram-bot==13.11
python-dotenv==1.0.0
